requirements.txt
azure-functions

azure-core==1.7.0

azure-storage-blob==12.3.2

certifi==2020.6.20

cffi==1.14.0

chardet==3.0.4

cryptography==2.9.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

isodate==0.6.0

msrest==0.6.17

numpy==1.19.0

oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

opencv-python==4.3.0.36

pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

requests-oauthlib==1.3.0

requests==2.24.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

urllib3==1.25.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'